+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+**[0.9.0] 2018-10-11**
+ - Added command line options (run `gkraken -h` for more info)
+ - Added command line option to start the app hidden
+ - Added command line option to add/remove udev rule
+
 **[0.8.0] 2018-10-09**
  - Improved fan and pump charts
  - Better driver error handling